+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 29.0-SNAPSHOT - unreleased
+## 28.1.0 - 2025-02-13
 
 ### 🎁 New Features